refactor: 메뉴 그룹과 관련된 도메인을 별도의 패키지로 분리
