+++ src/main/java/kitchenpos/application/MenuService.java
@@ -5,10 +5,10 @@ import java.util.List;
 import kitchenpos.application.exception.MenuGroupNotFoundException;
 import kitchenpos.application.exception.ProductNotFoundException;
 import kitchenpos.domain.menu.repository.MenuRepository;
-import kitchenpos.repository.MenuGroupRepository;
+import kitchenpos.domain.menugroup.repository.MenuGroupRepository;
 import kitchenpos.domain.product.repository.ProductRepository;
 import kitchenpos.domain.menu.Menu;
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.domain.menugroup.MenuGroup;
 import kitchenpos.domain.menu.MenuProduct;
 import kitchenpos.domain.product.Product;
 import kitchenpos.ui.dto.request.CreateMenuProductRequest;
