+++ src/main/java/kitchenpos/ui/dto/response/ReadMenuGroupResponse.java
@@ -1,6 +1,6 @@
 package kitchenpos.ui.dto.response;
 
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.domain.menugroup.MenuGroup;
 
 public class ReadMenuGroupResponse {
