+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -9,11 +9,11 @@ import java.util.List;
 import kitchenpos.application.exception.MenuGroupNotFoundException;
 import kitchenpos.config.IntegrationTest;
 import kitchenpos.domain.menu.Menu;
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.domain.menugroup.MenuGroup;
 import kitchenpos.domain.menu.MenuProduct;
 import kitchenpos.domain.product.Product;
 import kitchenpos.domain.exception.InvalidPriceException;
-import kitchenpos.repository.MenuGroupRepository;
+import kitchenpos.domain.menugroup.repository.MenuGroupRepository;
 import kitchenpos.domain.menu.repository.MenuRepository;
 import kitchenpos.domain.product.repository.ProductRepository;
 import kitchenpos.ui.dto.request.CreateMenuProductRequest;
