+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -11,7 +11,7 @@ import java.util.Collections;
 import java.util.List;
 import kitchenpos.config.IntegrationTest;
 import kitchenpos.domain.menu.Menu;
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.domain.menugroup.MenuGroup;
 import kitchenpos.domain.menu.MenuProduct;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderLineItem;
@@ -21,7 +21,7 @@ import kitchenpos.domain.product.Product;
 import kitchenpos.domain.TableGroup;
 import kitchenpos.domain.exception.InvalidOrderStatusCompletionException;
 import kitchenpos.domain.exception.InvalidOrderTableSizeException;
-import kitchenpos.repository.MenuGroupRepository;
+import kitchenpos.domain.menugroup.repository.MenuGroupRepository;
 import kitchenpos.domain.menu.repository.MenuRepository;
 import kitchenpos.repository.OrderRepository;
 import kitchenpos.repository.OrderTableRepository;
