+++ src/test/java/kitchenpos/domain/menu/MenuProductTest.java
@@ -6,7 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.math.BigDecimal;
 import java.util.List;
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.domain.menugroup.MenuGroup;
 import kitchenpos.domain.product.Product;
 import kitchenpos.domain.exception.InvalidQuantityException;
 import org.junit.jupiter.api.DisplayNameGeneration;
